Card stylying/ Some issue fixes
Gave cards a color boarder based on the subject it is. Looks better. Fixed an issue where when user likes a post it sends them back to the first 1-5 posts, need to fix the ame issue for when they click to comment on a post. Keeps sending to first 1-5 posts